+++ examples/design-method-simulateDualResponsesSamplesDesign.R
@@ -48,10 +48,8 @@ design <- DualResponsesSamplesDesign(nextBest=mynextbest,
                                      increments=myIncrements)
 ##specified the true DLE and efficacy curve
 myTruthDLE <- probFunction(DLEmodel, phi1 = -53.66584, phi2 = 10.50499)
+myTruthEff <- efficacyFunction(Effmodel, theta1 = -4.818429, theta2 = 3.653058)
 
-myTruthEff<- function(dose)
-{Effmodel@ExpEff(dose,theta1=-4.818429,theta2=3.653058)
-}
 ##The true gain curve can also be seen
 myTruthGain <- function(dose)
 {return((myTruthEff(dose))/(1+(myTruthDLE(dose)/(1-myTruthDLE(dose)))))}